feat: add refreshToken and accessToken to the returned user object

+++ routes/refresh.ts
@@ -13,7 +13,7 @@ export default eventHandler(async (event) => {
     throw createError({ message: 'User not found!', status: 404 })
   }
 
-  const { save, deleteByUserId } = useTokens({
+  const { save, deleteByUserId, refreshToken, accessToken } = useTokens({
     event,
     userId,
     email: user.email!,
@@ -22,5 +22,9 @@ export default eventHandler(async (event) => {
   await deleteByUserId()
   await save()
 
-  return user
+  return {
+    ...user,
+    refreshToken,
+    accessToken
+  }
 })
